Implemented handler and connection to controller, no sanitizing of inputs yet.

+++ KtorServer/src/main/kotlin/com/example/paths/CompetitionsPath.kt
@@ -5,7 +5,7 @@ import io.ktor.locations.*
 @Location("/competitions")
 data class CompetitionsPath(
     val userId: Int = -1,
-    val timestamp: Int = 0,
+    val timestamp: Long = 0,
 ) {
     @Location("/{id}")
     data class CompetitionId(private val parent: CompetitionsPath, val id: Int = -1)
